test: migrate CollectionAlgorithmDecision to vitest
